requirements.txt
streamlit==1.11.1
openpyxl==3.0.10
plotly-express==0.4.1
pandas==1.4.3
Pillow==9.2.0
